refactor: change structure to follow DDD
