+++ .eslintrc.json
@@ -21,6 +21,12 @@
       {
         "ts": "never"
       }
+    ],
+    "@typescript-eslint/interface-name-prefix": [
+      "error",
+      {
+        "prefixWithI": "always"
+      }
     ]
   },
   "settings": {
