+++ src/modules/appointments/infra/http/routes/appointments.routes.ts
@@ -2,10 +2,10 @@ import { Router } from 'express'
 import { getCustomRepository } from 'typeorm'
 import { parseISO } from 'date-fns'
 
-import AppointmentsRepository from '../repositories/AppointmentsRepository'
-import CreateAppointmentService from '../services/CreateAppointmentService'
+import AppointmentsRepository from '@modules/appointments/repositories/AppointmentsRepository'
+import CreateAppointmentService from '@modules/appointments/services/CreateAppointmentService'
 
-import ensureAuthenticated from '../middleware/ensureAuthenticated'
+import ensureAuthenticated from '@modules/users/infra/http/middleware/ensureAuthenticated'
 
 const appointmentsRouter = Router()
