+++ src/modules/appointments/repositories/IAppointmentsRepository.ts
@@ -1,0 +1,7 @@
+import Appointment from '../infra/typeorm/entities/Appointment'
+
+interface IAppointmentsRepository {
+  findByDate(date: Date): Promise<Appointment | undefined>
+}
+
+export default IAppointmentsRepository
