+++ src/modules/appointments/services/CreateAppointmentService.ts
@@ -1,9 +1,10 @@
 import { startOfHour } from 'date-fns'
 import { getCustomRepository } from 'typeorm'
-import Appointment from '../models/Appointment'
-import AppointmentsRepository from '../repositories/AppointmentsRepository'
 
-import AppError from '../errors/AppError'
+import AppError from '@shared/errors/AppError'
+
+import Appointment from '../infra/typeorm/entities/Appointment'
+import AppointmentsRepository from '../repositories/AppointmentsRepository'
 
 interface Request {
   provider_id: string
