+++ src/modules/users/services/AuthenticateUserService.ts
@@ -2,10 +2,10 @@ import { getRepository } from 'typeorm'
 import { compare } from 'bcryptjs'
 import { sign, verify } from 'jsonwebtoken'
 
-import authConfig from '../config/auth'
-import AppError from '../errors/AppError'
+import authConfig from '@config/auth'
+import AppError from '@shared/errors/AppError'
 
-import User from '../models/User'
+import User from '../infra/typeorm/entities/User'
 
 interface Request {
   email: string
