--- src/routes/index.ts
@@ -1,12 +1,0 @@
-import { Router } from 'express'
-import appointmentsRouter from './appointments.routes'
-import usersRouter from './user.routes'
-import sessionsRouter from './sessions.routes'
-
-const routes = Router()
-
-routes.use('/appointments', appointmentsRouter)
-routes.use('/users', usersRouter)
-routes.use('/auth', sessionsRouter)
-
-export default routes
