+++ src/shared/infra/http/server.ts
@@ -5,10 +5,11 @@ import cors from 'cors'
 import 'express-async-errors'
 
 import routes from './routes'
-import './database'
 
-import uploadConfig from './config/upload'
-import AppError from './errors/AppError'
+import uploadConfig from '@config/upload'
+import AppError from '@shared/errors/AppError'
+
+import '@shared/infra/typeorm'
 
 const app = express()
